Extract narrative.js from gameLogic.js (closes #124)
